Make activity command top level
